policies for rolepermissions and test for course and rolepermissions ...

+++ modules/Hadihosseini88/Course/Http/Controllers/CourseController.php
@@ -48,9 +48,11 @@ class CourseController extends Controller
     public function update($id, CourseRequest $request, CourseRepo $courseRepo)
     {
         $course = $courseRepo->findByid($id);
+        $this->authorize('edit',$course);
         if ($request->hasFile('image')) {
-            $course->banner->delete();
             $request->request->add(['banner_id' => MediaFileService::upload($request->file('image'))->id]);
+            if ($course->banner)
+                $course->banner->delete();
         } else {
             $request->request->add(['banner_id'=>$course->banner_id]);
         }
@@ -62,7 +64,7 @@ class CourseController extends Controller
     public function destroy($id, CourseRepo $courseRepo)
     {
         $course = $courseRepo->findByid($id);
-
+        $this->authorize('delete',$course);
         if ($course->banner) {
             $course->banner->delete();
         }
@@ -74,12 +76,14 @@ class CourseController extends Controller
 
     public function accept($id, CourseRepo $courseRepo)
     {
+        $this->authorize('change_confirmation_status',Course::class);
         if ($courseRepo->updateConfirmationStatus($id,Course::CONFIRMATION_STATUS_ACCEPTED)){
             return AjaxResponses::SuccessResponse();
         }
         return AjaxResponses::FailedResponse();
     }
     public function reject($id, CourseRepo $courseRepo) {
+        $this->authorize('change_confirmation_status',Course::class);
         if ($courseRepo->updateConfirmationStatus($id,Course::CONFIRMATION_STATUS_REJECTED)){
             return AjaxResponses::SuccessResponse();
         }
@@ -87,6 +91,7 @@ class CourseController extends Controller
     }
 
     public function lock($id, CourseRepo $courseRepo) {
+        $this->authorize('change_confirmation_status',Course::class);
         if ($courseRepo->updateStatus($id,Course::STATUS_LOCKED)){
             return AjaxResponses::SuccessResponse();
         }
